Move checkAvailableDiskSpace() to pacman
