+++ src/libaktualizr/package_manager/packagemanagerinterface.h
@@ -51,6 +51,7 @@ class PackageManagerInterface {
   virtual bool fetchTarget(const Uptane::Target& target, Uptane::Fetcher& fetcher, const KeyManager& keys,
                            const FetcherProgressCb& progress_cb, const api::FlowControlToken* token);
   virtual TargetStatus verifyTarget(const Uptane::Target& target) const;
+  virtual bool checkAvailableDiskSpace(const uint64_t required_bytes) const;
 
  protected:
   PackageConfig config;
